GenAI/llamaindex/llamaindex-docs-agent/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
azure-ai-documentintelligence==1.0.0
azure-core==1.32.0
azure-identity==1.20.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cobble==0.1.4
coloredlogs==15.0.1
cryptography==44.0.1
dataclasses-json==0.6.7
defusedxml==0.7.1
deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
email-validator==2.2.0
et-xmlfile==2.0.0
fastapi==0.115.8
fastapi-cli==0.0.7
fastembed==0.5.1
ffmpy==0.5.0
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
gradio==5.12.0
gradio-client==1.5.4
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
isodate==0.7.2
jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
llama-cloud==0.1.13
llama-cloud-services==0.6.1
llama-index==0.12.19
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.0
llama-index-core==0.12.19
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.8
llama-index-llms-openai==0.3.20
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.5
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-qdrant
llama-parse==0.6.1
loguru==0.7.3
lxml==5.3.1
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==0.14.1
markitdown==0.0.1a4
markupsafe==2.1.5
marshmallow==3.26.1
mdurl==0.1.2
mmh3==4.1.0
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
olefile==0.47
onnxruntime==1.20.1
openai==1.63.2
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathvalidate==3.2.3
pdfitdown==0.0.5
pdfminer-six==20240706
pillow==10.4.0
portalocker==2.10.1
propcache==0.3.0
protobuf==5.29.3
puremagic==1.28
py-rust-stemmers==0.1.5
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydub==0.25.1
pygments==2.19.1
pyjwt==2.10.1
pymupdf==1.25.1
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.1
pyyaml==6.0.2
qdrant-client==1.13.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
ruff==0.9.7
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
speechrecognition==3.14.1
sqlalchemy==2.0.38
starlette==0.45.3
striprtf==0.0.26
sympy==1.13.3
tavily-python==0.5.1
tenacity==9.0.0
termcolor==2.5.0
tiktoken==0.9.0
tokenizers==0.21.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.1
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.1
urllib3==2.3.0
uv==0.6.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2
wrapt==1.17.2
xlrd==2.0.1
xlsxwriter==3.2.2
yarl==1.18.3
youtube-transcript-api==0.6.3